Change projects widgets imports after moving.
