+++ spyder/plugins/projects/widgets/configdialog.py
@@ -13,8 +13,8 @@ from spyder.config.base import _
 from spyder.preferences.configdialog import ConfigDialog, GeneralConfigPage
 from spyder.utils.qthelpers import get_icon
 from spyder.config.user import NoDefault
-from spyder.widgets.projects import EmptyProject
-from spyder.widgets.projects.config import (WORKSPACE, VCS, ENCODING,
+from spyder.plugins.projects.widgets import EmptyProject
+from spyder.plugins.projects.widgets.config import (WORKSPACE, VCS, ENCODING,
                                             CODESTYLE)
