+++ spyder/plugins/projects/widgets/type/python.py
@@ -11,7 +11,7 @@ import os
 import os.path as osp
 
 from spyder.config.base import _
-from spyder.widgets.projects.type import EmptyProject
+from spyder.plugins.projects.widgets.type import EmptyProject
 
 
 class PythonProject(EmptyProject):
